Separate librustc module
